fixed a bug in the ComponentManager where minified builds would not map components to the correct key making getComponent always fail
In the code, when a component is added to the ComponentManager, the constructor.name of the component is used as key for each list of components. This is a problem, as constructor.name is different when the build is minified (as classes names change), so when calling the getComponent method with the component name as a string, the strings (e.g. "Transform") do not match the keys (e.g. "e") anymore.